requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
elastic-transport==8.15.1
elasticsearch==8.15.1
idna==3.10
jsons==1.6.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
typish==1.9.3
urllib3==2.2.3
